+++ resources/views/livewire/organizations/brands/branches/index.blade.php
@@ -120,7 +120,7 @@
                             @endif
                         </div>
 
-                        @if ($canManageBranches || $canManageZones || $canChangeServicePointStatus || $canGenerateQr || $canManageStaff)
+                        @if ($canManageBranches || $canManageZones || $canChangeServicePointStatus || $canOpenTable || $canGenerateQr || $canManageStaff)
                             <div class="flex flex-wrap gap-2 md:justify-end">
                                 @if ($canManageZones)
                                     <flux:button icon="rectangle-group" type="button" :href="route('organizations.brands.branches.areas.index', [$organization, $brand, $branch])" wire:navigate>
@@ -128,7 +128,7 @@
                                     </flux:button>
                                 @endif
 
-                                @if ($canChangeServicePointStatus || $canGenerateQr)
+                                @if ($canChangeServicePointStatus || $canOpenTable || $canGenerateQr)
                                     <flux:button icon="squares-2x2" type="button" :href="route('organizations.brands.branches.service-points.index', [$organization, $brand, $branch])" wire:navigate>
                                         {{ __('Service points') }}
                                     </flux:button>